Created basic model package
Created the classes depending on user and the class User; Course, Moment, StudySession, ToDo. I have also implemented important methods and variables. Mostly made methods so that the classes will work and be settable. All of the now discussed variables are also implemented. Everything is for now set to private.