Have completed the functionality of the sidePanel
All buttons now function and the app initialises with the coursePage on top. Instead of loading a controller with and fxml it now loads an fxml file with an controller. A factory has been implemented to further the struction and abstraction between the modules. It is now cleanly build and open to extension. The application now created three pages that are put in the controller and are obtained in the sidePanel that decides witch ones are to be shown.

Also some classes have been renamed to fit better