Have completed the Userstory
Added to the Factory so that now all FXML can be loaded and created here. Now all the courses show up in a clickable list where another anchorpane is pressed to front when we press the button. If we now press courses the page will "reset" and have the list in front. A lot of changes have been made especially in the CourseSelectionPage. Also the FXMLs have bee developed and set to fit these new changes.