Fixed Course- editing, deletion and status change
Added fxml and functionality so the user now can edit the name of the course and the code. The user can also delete the course if wished with a second chance to change their mind. The User can also complete the course and also reactivate the course.

All of these changes are notified to all the listeners of the CourseManager.

Also deleted previous fault from merge