database: implement a basic database
This initial commit implements basic database operations (only read
operations so far) for users and courses.

There was some issues with setting the currently logged in user via the
dependency injection container. The current workaround is to listen to
UserChangedEvent and pass the user object around to the classes and
methods.